Swweb 2994 (#1718)
* oracle initial impl

* fixing commlog queries

* adding missing file

* fixing dashboards

* restoring missing dlls

* fixing dlls declarations

* fixing deps++

* fixing searches for oracle

* adding missing nuget

* fixing tests

* deleting files that got added

* updating nuget packages

* activity stream ++

* fixing offline

* fixing extra oracle queries

+++ softWrench.sW4.Web/DB_Migration/5.19.0/20170210_SWWEB_2918.cs
@@ -1,5 +1,6 @@
 ﻿using cts.commons.persistence.Util;
 using FluentMigrator;
+using softwrench.sw4.api.classes.migration;
 
 namespace softWrench.sW4.Web.DB_Migration._5._19._0 {
     [Migration(201702101036)]
@@ -7,12 +8,15 @@ namespace softWrench.sW4.Web.DB_Migration._5._19._0 {
     public class MigrationSwweb2918 : Migration {
 
         public override void Up() {
-            Execute.Sql("delete from conf_propertydefinition where key_ = 'whereclause' and defaultvalue = '' and fullkey not in (select definition_id from conf_propertyvalue)");
+            if (!MigrationContext.IsOracle) {
+                Execute.Sql("delete from conf_propertydefinition where key_ = 'whereclause' and defaultvalue = '' and fullkey not in (select definition_id from conf_propertyvalue)");
+            }
+
             Alter.Table("CONF_PROPERTYVALUE").AddColumn("clientname").AsString(MigrationUtil.StringSmall).Nullable();
         }
 
         public override void Down() {
-            
+
         }
     }
 }
